storage and token formats support (legacy, opaque, JWT)

+++ example/adapters/redis.js
@@ -5,6 +5,12 @@ const client = new Redis(process.env.REDIS_URL, {
   keyPrefix: 'oidc:',
 });
 
+const grantable = [
+  'AccessToken',
+  'AuthorizationCode',
+  'RefreshToken',
+];
+
 function grantKeyFor(id) {
   return `grant:${id}`;
 }
@@ -14,17 +20,57 @@ class RedisAdapter {
     this.name = name;
   }
 
-  key(id) {
-    return `${this.name}:${id}`;
+  upsert(id, payload, expiresIn) {
+    const key = this.key(id);
+    const store = grantable.includes(this.name) ? {
+      dump: JSON.stringify(payload),
+      ...(payload.grantId ? { grantId: payload.grantId } : undefined),
+    } : JSON.stringify(payload);
+
+    const multi = client.multi();
+    multi[grantable.includes(this.name) ? 'hmset' : 'set'](key, store);
+
+    if (expiresIn) {
+      multi.expire(key, expiresIn);
+    }
+
+    if (payload.grantId) {
+      const grantKey = grantKeyFor(payload.grantId);
+      multi.rpush(grantKey, key);
+    }
+
+    return multi.exec();
+  }
+
+  async find(id) {
+    const data = grantable.includes(this.name) ?
+      await client.hgetall(this.key(id)) :
+      await client.get(this.key(id));
+
+    if (isEmpty(data)) {
+      return undefined;
+    }
+
+    if (typeof data === 'string') {
+      return JSON.parse(data);
+    }
+    const { dump, ...rest } = data;
+    return {
+      ...rest,
+      ...JSON.parse(dump),
+    };
   }
 
   async destroy(id) {
     const key = this.key(id);
-    const grantId = await client.hget(key, 'grantId');
-    const tokens = await client.lrange(grantKeyFor(grantId), 0, -1);
-    const deletions = tokens.map(token => client.del(token));
+    const deletions = [];
+    if (grantable.includes(this.name)) {
+      const grantId = await client.hget(key, 'grantId');
+      const tokens = await client.lrange(grantKeyFor(grantId), 0, -1);
+      tokens.forEach(token => deletions.push(client.del(token)));
+      deletions.push(client.del(grantKeyFor(grantId)));
+    }
     deletions.push(client.del(key));
-    deletions.push(client.del(grantKeyFor(grantId)));
     await deletions;
   }
 
@@ -32,40 +78,8 @@ class RedisAdapter {
     return client.hset(this.key(id), 'consumed', Math.floor(Date.now() / 1000));
   }
 
-  async find(id) {
-    const data = await client.hgetall(this.key(id));
-    if (isEmpty(data)) {
-      return undefined;
-    } else if (data.dump !== undefined) {
-      return JSON.parse(data.dump);
-    }
-    return data;
-  }
-
-  upsert(id, payload, expiresIn) {
-    const key = this.key(id);
-    let toStore = payload;
-
-    // Clients are not simple objects where value is always a string
-    // redis does only allow string values =>
-    // work around it to keep the adapter interface simple
-    if (this.name === 'Client' || this.name === 'Session') {
-      toStore = { dump: JSON.stringify(payload) };
-    }
-
-    const multi = client.multi();
-    multi.hmset(key, toStore);
-
-    if (expiresIn) {
-      multi.expire(key, expiresIn);
-    }
-
-    if (toStore.grantId) {
-      const grantKey = grantKeyFor(toStore.grantId);
-      multi.rpush(grantKey, key);
-    }
-
-    return multi.exec();
+  key(id) {
+    return `${this.name}:${id}`;
   }
 }
